Fixed reporting error and better logging
The processor tried to set the success status, but this caused an exception because `_success_status` was declared with a type but not a value. The error was hidden and not logged because `log` wasn't getting updated after the _process(). Updating the entire log array is difficult because multiple reports might be generated at the same time and we could overwrite changes from another one, so we need to switch to the list_append_fields approach.