Lots of updates (#9)
* Random UI changes

* Fixed reporting error and better logging

The processor tried to set the success status, but this caused an exception because `_success_status` was declared with a type but not a value. The error was hidden and not logged because `log` wasn't getting updated after the _process(). Updating the entire log array is difficult because multiple reports might be generated at the same time and we could overwrite changes from another one, so we need to switch to the list_append_fields approach.

* Detect changes to the file

Don't force components to react if file doesn't change. Also, provide a note that the records list might need to be refreshed.

* Other changes

* Refactor records table and add value columns

* More updates

* Exceptions and exception handling

Add consistency and more metadata to the 1st-party exceptions (now instances of RivoliError). Added testing.

* Record Processor Exception Logging

Create Rivoli Logs from Exceptions in a consistent way with as much exception-related metadata as possible.

* API Request Logging

Plus tests

* Loader: Fixed Width Support

* Bug: Don't upload failed records

* Add report filtering based on recently-failed function ids

TODO: Add chooser support in the UI. This will need to have a dynamic list of used validation functions, and also handle situations where a function is removed. Should also include copious validation.

* UI Tweaks

* Refactor Exceptions

Move auto_retry and http_response from ExecutionError to ConfigurationError

* Fix: Error with property not found

Move the max_pending_updates to the parent RecordProcessor for the Loader changes a few CLs ago.

* WIP on fixed-width file processing

* String Components

Moved away from the auto-add and auto-remove model of new items. Added an "add" button. This greatly simplifies the management and mental model of these.
Looks like the remove buttons were never enabled, so those were implemented.

* Add column of filetypes to the Partners list (table)

* Added Notification component

Shows a toast notification after ajax actions rather than using alert()

* Fixed Function Editors

Tried to fix issues with reactivity, searching, and situations where the selected function has been deactivated (but still needs to be shown).

* Record Searching

* More notifications

* Add bson formatting functions and tests

To support future commit in which the Reporter has to update specific items within a document to prevent overwriting the rest of the document.

* Reporter targets db updates to only relevant sub-docs

These changes prevent race conditions due to reporter updating too much of a File doc in the db.

Breaking change due to trying to target the individual step stats for filtering and updating since they had .'s in their key.

* raise_validation_error() wrapper

Similar to raise_configuration_error()

* Protobuf WKT Timestamp Support

Only supports converting from a timestamp field to a datetime instance and not the other way around.

* Add timestamp field to apilog message

Eventually this will be used to set a TTL index on the apilog collection

* Scripts to generate CSV file and record files

These can be loaded into a local sqlite db (for which there's a helper script) or into something like snowflake.

* Allow reports before file is complete

No mechanism to auto-generate reports at this point, but they can be manually generated.

* Support empty report filename templates

If the configured template is empty then create a filename based on the report name

* Better formatting of "no items" message.

* Support processing diagrams without processed files

This will be used for File configuration where the diagram is shown but without a processed file

* Fix bug with showing notification in UI

And cleanup some related code

* Better ApiLog support

* Moving toward enum-based levels to determine when to write the logs. When will this be set, though? Runtime, or at the file-level?
* Better text overflow clipping

* Catch all request exceptions

Move the json() decoding into the try block so that the except block can catch any errors. Changed the except filter to the RequestException, which is a parent of all previous except block filtered exceptions, plus JSONDecodeError, plus lots more.

Not catching standard python exceptions because those would be unexpected here.

* Removed delimiter check

The CSV sniffer was acting up and detected the wrong delimiter, which prevented any file processing. Changed the exception to a warning log message and began using the specified delimiter instead of the detected delimiter.

* Fix notification popup where message is not provided

* Renamed *_keys to fields_*

input_keys -> fields_in & output_keys -> fields_out

These will (soon) be used to specify the expected input/output fields of the Record dict.

* Add enum support to functions

Enums as function parameters are a way of specifying (and enforcing) a list of possible values.

* Modification & Validation Typing

Function types (e.g., FIELD_VALIDATION, RECORD_VALIDATION) have varying input/output types. For example, a FIELD_VALIDATION function should always return the value of that particular field while RECORD_VALIDATION can return a dict, a Record, or None. We define all of that here.

* Bug fixes

* Notes about exception types

* Add any Api log id to the exception

* Fix: don't validate records which have already been validated

* Make field values in records table copiable

* Support fixed-length fields in UI

* Cleanup and comments

* Fixed with support

* WIP - Add processing diagram preview to file config page

On clicking a button we should render the processing diagram for visualization purposes

* Add failed function support to OutputSubForm

Failed function IDs are already supported by the report generator -- this exposes the configuration in the UI

* Updated gitignore files

* REcordTypeSubForm updates

* Convert accordion to a database
* Support display of fixed-width field ranges

* Updated node packages

* API Exception Summarization

Add a function to clean exception messages coming from requests. This attempts to remove some high-cardinality stuff (specifically the URL) to create a cleaner summary string.