Feat: Check if record exists earlier in file
The checking of hashes of previously UPLOADED records works on the chunk for efficiency purposes (so that it can query, e.g., 1000 hashes from mongo in a single query rather than make 1000 queries). The theoretical problem is that it won't detect a collision with a record that hasn't been uploaded yet (ie, one that is earlier in the chunk and will be uploaded). This change stores the "chunk hashes" and checks against those as well.

TODO:
* Report the record ID / line number of the collision (both for previously-uploaded and in-file collisions)
* Copy this functionality to the validation step, or at least create a validator for this. Don't remove it from the UPLOAD stage because there should be a final check just before upload.