Docker - Processor
Create a processor image in docker, run it with compose

The build script copies in any third_party modules; this is specifically for the inclusion of third_party functions.

The container has a defined volume to store its data files (specifically the dropped and processed files and the report outputs. This volume will also be shared with the webui which needs to upload files.

The image will be run in two different instances -- as a standard celery worker and as a celery beat scheduler. This is best practice as per celery docs.