Refactoring DB access approach
DB client is only created when get_db() is called, and now modules only call get_db() in code, rather than at the top of the module. The goal is to allow containers which don't need to access the DB (such as the schedule, or maybe a future file-watcher) to not require a mongo connection (or credentials).