Create separate input directory configuration
Previously (and now by default) the input directory was a subdirectory of the main files directory. Now it's separately configurable, which allows for the main files directory to be a container volume and the input directory to be bound to the host. (The input files are scanned, matching files are moved into the main directory structured, then removed from the input directory.)

The docker compose file has been set up to bind this directory to ~/Documents/rivoli/input_files and the build script creates this directory.